stuck on collision problems
- as AreaObjects get larger, their collision region gets more inaccurate to where they're rendered. wasted about 8 hours on this problem so far
- going to stop trying to make the current algorithm work and start moving to a different pattern

+++ src/net/wforbes/omnia/overworld/entity/Entity.java
@@ -16,7 +16,7 @@ public abstract class Entity implements Renderable {
     public abstract double getY();
     public abstract String getName();
     public abstract int getCollisionBoxWidth();
-    public abstract int getCollisionRadius();
+    public abstract double getCollisionRadius();
 
     public abstract void init();
     public abstract void init(int xPos, int yPos);
